backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-extra-types[email]==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.2
markdown==3.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 